@@ -5,14 +5,11 @@ import Quartz
 // MARK: - CGWindowClient
 
 @DependencyClient
-public struct CGWindowClient {
-  public var list: (
-    _ options: CGWindowListOption /* [.excludeDesktopElements, .optionOnScreenOnly] */,
-    _ referenceWindow: CGWindowID /* kCGNullWindowID */ ) -> [CGWindowValue] = { _, _ in [] }
+public struct CGWindowClient: Sendable {
+  public var list: @Sendable (_ options: CGWindowListOption, _ referenceWindow: CGWindowID) -> [CGWindowValue] = { _, _ in [] }
   // https://github.com/nonstrict-hq/ScreenCaptureKit-Recording-example/blob/main/Sources/sckrecording/main.swift
-  public var preflightScreenCaptureAccess: () -> Bool = { false }
-  public var requestScreenCaptureAccess: () -> Bool = { false }
-
+  public var preflightScreenCaptureAccess: @Sendable () -> Bool = { false }
+  public var requestScreenCaptureAccess: @Sendable () -> Bool = { false }
 }
 
 // MARK: DependencyKey
@@ -20,10 +17,11 @@ public struct CGWindowClient {
 extension CGWindowClient: DependencyKey {
   public static let liveValue = Self(
     list: { options, referenceWindow in
+      // CGWindowListCopyWindowInfo([.excludeDesktopElements, .optionOnScreenOnly], kCGNullWindowID)
       (CGWindowListCopyWindowInfo(options, referenceWindow) as? [[CFString: Any]])?.compactMap(CGWindowValue.init) ?? []
     },
-    preflightScreenCaptureAccess: CGPreflightScreenCaptureAccess,
-    requestScreenCaptureAccess: CGRequestScreenCaptureAccess)
+    preflightScreenCaptureAccess: { CGPreflightScreenCaptureAccess() },
+    requestScreenCaptureAccess: { CGRequestScreenCaptureAccess() })
 
   public static let testValue = Self()
 }
