@@ -5,15 +5,15 @@ import Foundation
 // MARK: - FileManagerClient
 
 @DependencyClient
-public struct FileManagerClient {
-  public var urls: (
+public struct FileManagerClient: Sendable {
+  public var urls: @Sendable (
     _ directory: FileManager.SearchPathDirectory,
     _ domainMask: FileManager.SearchPathDomainMask) -> [URL] = { _, _ in [] }
-  public var contentsOfDirectory: (
+  public var contentsOfDirectory: @Sendable (
     _ url: URL,
     _ keys: [URLResourceKey]?,
-    _ mask: FileManager.DirectoryEnumerationOptions) throws -> [URL]
-  public var createDirectory: (
+    _ options: FileManager.DirectoryEnumerationOptions) throws -> [URL]
+  public var createDirectory: @Sendable (
     _ atURL: URL,
     _ withIntermediateDirectories: Bool,
     _ attributes: [FileAttributeKey: Any]?) throws
@@ -24,9 +24,9 @@ public struct FileManagerClient {
 
 extension FileManagerClient: DependencyKey {
   public static let liveValue = FileManagerClient(
-    urls: FileManager.default.urls,
-    contentsOfDirectory: FileManager.default.contentsOfDirectory,
-    createDirectory: FileManager.default.createDirectory)
+    urls: { FileManager.default.urls(for: $0, in: $1) },
+    contentsOfDirectory: { try FileManager.default.contentsOfDirectory(at: $0, includingPropertiesForKeys: $1, options: $2) },
+    createDirectory: { try FileManager.default.createDirectory(at: $0, withIntermediateDirectories: $1, attributes: $2) })
   public static let testValue = FileManagerClient()
 }
 
