@@ -5,12 +5,12 @@ import DependenciesMacros
 // MARK: - NSScreenClient
 
 @DependencyClient
-public struct NSScreenClient {
+public struct NSScreenClient: Sendable {
   // alt-tab's handling of `NSScreen.main`: https://github.com/lwouis/alt-tab-macos/blob/8152c3cebf0091385d8ec8d54aeefd805477d864/src/logic/NSScreen.swift#L28
   // see also -- NSScreen.main is not really the active screen: https://stackoverflow.com/a/56268826
-  public var main: () -> NSScreenValue?
-  public var deepest: () -> NSScreenValue?
-  public var screens: () -> [NSScreenValue] = { [] }
+  public var main: @Sendable () -> NSScreenValue?
+  public var deepest: @Sendable () -> NSScreenValue?
+  public var screens: @Sendable () -> [NSScreenValue] = { [] }
 
   public static let liveValue = NSScreenClient(
     main: { NSScreen.main.map(NSScreenValue.init) },
