@@ -5,14 +5,16 @@ import Foundation
 // MARK: - URLSessionClient
 
 @DependencyClient
-public struct URLSessionClient {
-  public var data: (_ request: URLRequest) async throws -> (Data, URLResponse)
+public struct URLSessionClient: Sendable {
+  public var data: @Sendable (_ request: URLRequest) async throws -> (Data, URLResponse)
 }
 
 // MARK: DependencyKey
 
 extension URLSessionClient: DependencyKey {
-  public static let liveValue: Self = .init(data: URLSession.shared.data)
+  public static let liveValue = Self(
+    data: { try await URLSession.shared.data(for: $0) }
+  )
   public static let testValue = Self()
 }
 
