@@ -1,6 +1,6 @@
 import Foundation
 
-extension NSKeyValueChange: CustomDebugStringConvertible {
+extension NSKeyValueChange: @retroactive CustomDebugStringConvertible {
   public var debugDescription: String {
     switch self {
     case .insertion:
