@@ -12,7 +12,7 @@ extension NSMenu {
 
   // MARK: Public
 
-  public static func fromNib(named nibName: String, bundle: Bundle = .main) -> NSMenu {
+  @MainActor public static func fromNib(named nibName: String, bundle: Bundle = .main) -> NSMenu {
     let nib = NSNib(nibNamed: nibName, bundle: bundle)
     var objects: NSArray?
     nib?.instantiate(withOwner: NSApplication.shared.delegate, topLevelObjects: &objects)
@@ -23,7 +23,7 @@ extension NSMenu {
   // MARK: Internal
 
   /// A programmatically created application menu bar (`NSApp.mainMenu`). The menu’s structure and contents are an exact clone of those in a Storyboard–based, non–modified menu bar (as of **Xcode 14.2**). This is not meant to be used directly in your project (hence the lack of the `public` modifier), but rather copied and modified to suit your needs.
-  static let standardMenuBar = NSMenu(APP_NAME) {
+  @MainActor static let standardMenuBar = NSMenu(APP_NAME) {
     NSMenuItem(APP_NAME) {
       NSMenuItem("About \(APP_NAME)", action: #selector(NSApplication.orderFrontStandardAboutPanel(_:)))
       NSMenuItem.separator()
